Make code in eval.cc callable from other files
Context: We need to import some changes which add MLPerf logging to MiniGp.
Some of the MLPerf logging was added to eval_models.cc, which calls the
code in eval.cc. As a first step, we need to make the code in eval.cc
callable from other files. We do so by moving the Evaluator class into
its own evaluator.{h,cc}. The previous eval.cc then includes evaluator.h
to use Evaluator.

Tested:
bazel build cc/evaluator
bazel build cc/eval
