+++ reinforcement/tensorflow/minigo/cc/BUILD
@@ -485,10 +485,10 @@ minigo_cc_binary(
     ],
 )
 
-minigo_cc_binary(
-    name = "eval",
-    srcs = ["eval.cc"],
-    visibility = ["//visibility:public"],
+minigo_cc_library(
+    name = "evaluator",
+    srcs = ["evaluator.cc"],
+    hdrs = ["evaluator.h"],
     deps = [
         ":base",
         ":game",
@@ -508,6 +508,15 @@ minigo_cc_binary(
         "@com_google_absl//absl/strings",
         "@com_google_absl//absl/strings:str_format",
         "@com_google_absl//absl/time",
+    ],
+)
+
+minigo_cc_binary(
+    name = "eval",
+    srcs = ["eval.cc"],
+    visibility = ["//visibility:public"],
+    deps = [
+        ":evaluator",
     ],
 )
